put list search items into its own component
