+++ src/views/List.jsx
@@ -1,33 +1,9 @@
-import { useState } from 'react';
-import { ListItem } from '../components';
-import { comparePurchaseUrgency } from '../api/firebase';
 import './List.css';
 import AddItemForm from '../components/AddItemForm';
 import ShareForm from '../components/ShareForm';
+import ListSearchItems from '../components/ListSearchItems';
 
-export function List({ data, isShoppingListLoading, listPath, listName }) {
-	const [input, setInput] = useState('');
-
-	const sortedItems = comparePurchaseUrgency(data);
-
-	const filteredItems = sortedItems.filter((item) => {
-		if (input) {
-			const lowerCaseItemName = item.name.toLowerCase();
-			return lowerCaseItemName.includes(input.toLowerCase());
-		} else {
-			return item;
-		}
-	});
-
-	function handleInputChange(e) {
-		e.preventDefault();
-		setInput(e.target.value);
-	}
-
-	function clearSearch() {
-		setInput('');
-	}
-
+export function List({ data, isShoppingListLoading, listPath }) {
 	if (!listPath) {
 		return (
 			<>
@@ -43,34 +19,9 @@ export function List({ data, isShoppingListLoading, listPath, listName }) {
 
 	return (
 		<>
-			<h2>{listName}</h2>
-			{data.length === 0 && (
-				<h2 data-testid="noItemsErrorMsg">You have no items in this list!</h2>
-			)}
-			{data.length !== 0 && (
-				<div className="searchInput">
-					<form action="" onSubmit={(e) => e.preventDefault()}>
-						<label htmlFor="searchItems">Search your shopping list: </label>
-						<input
-							onChange={handleInputChange}
-							id="searchItems"
-							value={input}
-							type="text"
-						/>
-					</form>
-					<button onClick={clearSearch}>Clear</button>
-				</div>
-			)}
 			<div>
-				{filteredItems.map((item) => (
-					<ListItem key={item.id} item={item} listPath={listPath} />
-				))}
+				<ListSearchItems listPath={listPath} data={data} />
 			</div>
-			{data.length > 0 && filteredItems.length === 0 && (
-				<div>
-					<p>No match found for that filter query.</p>
-				</div>
-			)}
 			<div>
 				<AddItemForm listPath={listPath} data={data} />
 			</div>
